requirements.txt
annotated-types==0.7.0
anyio==4.8.0
app==0.0.1
click==8.1.8
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.115.11
fastapi-sessions==0.3.2
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jose==1.0.0
MarkupSafe==3.0.2
middleware==1.2.3
mysql-connector-python==9.2.0
mysqlclient==2.2.7
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
